+++ polyfill-service/documentation/routes/home-element/tests/index.spec.js
@@ -1,0 +1,7 @@
+import HomeElement from '../index';
+
+describe('HomeElement class', function desc() {
+  it('should have a is attribute with value home-page', function it() {
+    expect(HomeElement.is).to.equal('home-page');
+  });
+});